Move pruneResolvedTree() into a separate file
